# [RFC] Improved KernelMixture::computePDF() performances