--- a/lib/src/Uncertainty/Distribution/openturns/Epanechnikov.hxx
+++ b/lib/src/Uncertainty/Distribution/openturns/Epanechnikov.hxx
@@ -61,12 +61,15 @@ public:
 
   /** Get the PDF of the distribution */
   using ContinuousDistribution::computePDF;
+  Scalar computePDF(const Scalar scalar) const;
   Scalar computePDF(const Point & point) const;
 
   /** Get the CDF of the distribution */
   using ContinuousDistribution::computeCDF;
+  Scalar computeCDF(const Scalar scalar) const;
   Scalar computeCDF(const Point & point) const;
   using ContinuousDistribution::computeComplementaryCDF;
+  Scalar computeComplementaryCDF(const Scalar scalar) const;
   Scalar computeComplementaryCDF(const Point & point) const;
 
   /** Get the PDFGradient of the distribution */