--- a/lib/src/Uncertainty/Distribution/openturns/Gamma.hxx
+++ b/lib/src/Uncertainty/Distribution/openturns/Gamma.hxx
@@ -71,14 +71,18 @@ public:
 
   /** Get the PDF of the distribution */
   using ContinuousDistribution::computePDF;
+  Scalar computePDF(const Scalar scalar) const;
   Scalar computePDF(const Point & point) const;
   using ContinuousDistribution::computeLogPDF;
+  Scalar computeLogPDF(const Scalar scalar) const;
   Scalar computeLogPDF(const Point & point) const;
 
   /** Get the CDF of the distribution */
   using ContinuousDistribution::computeCDF;
+  Scalar computeCDF(const Scalar scalar) const;
   Scalar computeCDF(const Point & point) const;
   using ContinuousDistribution::computeComplementaryCDF;
+  Scalar computeComplementaryCDF(const Scalar scalar) const;
   Scalar computeComplementaryCDF(const Point & point) const;
 
   /** Compute the entropy of the distribution */