--- a/lib/src/Uncertainty/Distribution/openturns/KernelMixture.hxx
+++ b/lib/src/Uncertainty/Distribution/openturns/KernelMixture.hxx
@@ -185,7 +185,7 @@ private:
   void computeNormalizationFactor();
 
   /** The kernel of the kernelMixture */
-  Pointer<DistributionImplementation>  p_kernel_;
+  Pointer<DistributionImplementation> p_kernel_;
 
   /** The bandwidth of the KernelMixture */
   Point bandwidth_;