--- a/lib/src/Uncertainty/Distribution/openturns/Triangular.hxx
+++ b/lib/src/Uncertainty/Distribution/openturns/Triangular.hxx
@@ -72,10 +72,12 @@ public:
 
   /** Get the PDF of the Triangular distribution */
   using ContinuousDistribution::computePDF;
+  Scalar computePDF(const Scalar scalar) const;
   Scalar computePDF(const Point & point) const;
 
   /** Get the CDF of the Triangular distribution */
   using ContinuousDistribution::computeCDF;
+  Scalar computeCDF(const Scalar scalar) const;
   Scalar computeCDF(const Point & point) const;
 
   /** Get the characteristic function of the distribution, i.e. phi(u) = E(exp(I*u*X)) */