… lib/src/Uncertainty/Distribution/KernelMixtureTmp.cxx
… lib/src/Uncertainty/Distribution/ProductDistribution.cxx
… lib/src/Uncertainty/Distribution/RandomMixture.cxx
… lib/src/Uncertainty/Distribution/openturns/ProductDistribution.hxx
… lib/test/t_GaussKronrod_std.cxx
… lib/test/t_GaussKronrod_std.expout