There are two main bottlenecks:
  * The first one is due to the creation and destruction of Counter at each call to getImplementation() for interface objects.
  * The second one is due to the creation and destruction of Point during the call to kernel_.computePDF()

A more generic answer should be provided (hence the RFC), but this one already makes a speedup ranging from 9x to 18x!
 
The benchmark I use is:
```
from openturns import *

ResourceMap.SetAsUnsignedInteger( "Distribution-MinimumVolumeLevelSetSamplingSize", 20000 )

ref = Mixture([Normal(2), Normal([1.0]*2, [0.5]*2, IdentityMatrix(2))])
alpha = 0.03
level_ref, beta_ref = ref.computeMinimumVolumeLevelSetWithThreshold(1.0 - alpha)
size = 10000
data = ref.getSample(size)
from time import time
for kernel in [Uniform(), Epanechnikov(), Triangular(), Normal()]:
    print("kernel=", kernel)
    t0 = time()
    reconstructed = KernelSmoothing(kernel).build(data)
    level_rec, beta_rec = reconstructed.computeMinimumVolumeLevelSetWithThreshold(1.0 - alpha)
    t1 = time()
    print("t=%.2f" % (t1 - t0), "s")
```
With master, the results are:
```
kernel= Uniform(a = -1, b = 1)
t=14.79 s
kernel= Epanechnikov()
t=15.07 s
kernel= Triangular(a = -1, m = 0, b = 1)
t=14.83 s
kernel= Normal(mu = 0, sigma = 1)
t=21.89 s
```
and with this PR:
```
kernel= Uniform(a = -1, b = 1)
t=0.80 s
kernel= Epanechnikov()
t=0.82 s
kernel= Triangular(a = -1, m = 0, b = 1)
t=0.90 s
kernel= Normal(mu = 0, sigma = 1)
t=2.23 s
```
I also provide two flame graphs to illustrate the bottleneck
![perf_master](https://user-images.githubusercontent.com/8220933/61911397-a6528a00-af37-11e9-8afa-e282d92ad358.png)
![perf_PR](https://user-images.githubusercontent.com/8220933/61911398-a6eb2080-af37-11e9-9f41-e20168afbc3b.png)