@jschueller 2019-07-29
Did you benchmark separately the impact of interface ojects and Point creation ?

By relying on std::shared_ptr alone I'm already shaving 8-10% on your benchmark.

So if the impact of the interface is the same, it would be more interesting to not do it specifically here but globally.

FYI, my patch in Pointer.hxx:
```
diff --git a/lib/src/Base/Common/openturns/Pointer.hxx b/lib/src/Base/Common/openturns/Pointer.hxx
index ef35c7aae..b66db88c6 100644
--- a/lib/src/Base/Common/openturns/Pointer.hxx
+++ b/lib/src/Base/Common/openturns/Pointer.hxx
@@ -23,6 +23,7 @@
 
 #include "openturns/OTprivate.hxx"
 #include "openturns/SharedPointer.hxx"
+#include <memory>
 
 BEGIN_NAMESPACE_OPENTURNS
 
@@ -65,7 +66,7 @@ public:
 #ifdef SWIG
   typedef SharedPointer<T> pointer_type;
 #else
-  typedef IMPL<T> pointer_type;
+  typedef std::shared_ptr<T> pointer_type;
 #endif
 
   /**
@@ -136,7 +137,9 @@ public:
   template <class Base>
   Pointer & assign(const Pointer<Base> & ref)
   {
-    ptr_.assign(ref.ptr_);
+//     ptr_.assign(ref.ptr_);
+    std::cout << "Pointer::assign"<<std::endl;
+//     ptr_ = std::make_shared(ref);
     return *this;
   }
```