@regislebrun 2019-08-29
@jschueller Hi I am back to business. I did some additional benchmarks. With this branch and the old shared pointers, I get:
```
kernel= Uniform(a = -1, b = 1)
t=0.81 s
kernel= Epanechnikov()
t=0.83 s
kernel= Triangular(a = -1, m = 0, b = 1)
t=0.93 s
kernel= Normal(mu = 0, sigma = 1)
t=2.26 s
```
With the current master I get:
```
kernel= Uniform(a = -1, b = 1)
t=12.38 s
kernel= Epanechnikov()
t=13.43 s
kernel= Triangular(a = -1, m = 0, b = 1)
t=14.41 s
kernel= Normal(mu = 0, sigma = 1)
t=17.97 s
```
With both I get:
```
kernel= Uniform(a = -1, b = 1)
t=0.83 s
kernel= Epanechnikov()
t=0.82 s
kernel= Triangular(a = -1, m = 0, b = 1)
t=0.94 s
kernel= Normal(mu = 0, sigma = 1)
t=2.23 s
```
So the main improvement is mainly due to the avoidance of the Counter manipulation, and this PR leads to a significant improvement over master.