@jschueller 2019-08-29
ok, I expected a bit more magic from the cxx11 changes thought : |