@regislebrun 2019-09-03
I made additional speed improvements. Consider the following benchmark:
```
import openturns as ot
from time import time

size = 1000
d1 = ot.Normal()
d2 = ot.Normal()
p = d1 * d2
s = p.getSample(size)
t0 = time()
res = p.computePDF(s)
t1 = time()
print("Speed=", size / (t1 - t0), "pdf/s")
```
With master I get:
```
Speed= 5336.597115980623 pdf/s
```
By replacing Distribution by Pointer<DistributionImplementation> I get:
```
Speed= 6086.90253595881 pdf/s
```
Then by implementing the PDF kernel as a Function instead of a bound method I get:
```
 Speed= 9663.771662133502 pdf/s
```
There are still a lot of Point created/destroyed. I extended the GaussKronrod class to work with UniVariateFunction models and implemented the PDF kernel as a UniVariateFunction, and then I got:
```
Speed= 118513.26085353314 pdf/s
```
so a final speed-up of 22x!
Replacing Normal() by Beta(), master gives:
```
Speed= 10082.341762180316 pdf/s
```
and the various improvements already described give:
```
Speed= 37257.71140530983 pdf/s
```
so a speed-up of 3.7, not bad but not so impressive. It is due to the fact that there is no specific implementation of Scalar computePDF(const Scalar scalar), so the algorithm calls Scalar computePDF(const Point & point), which creates a lot of temporary objects. I implemented the method and got:
```
Speed= 89404.7967078228 pdf/s
```
now the speed-up is nearly 9x.

IMHO, the following points should be addressed:
- every distribution with 1D version (either pure 1D such as Gamma, or with an 1D version such as Dirichlet) should have specific implementations of computePDF(), computeLogPDF(), computeCDF(), computeComplementaryCDF() based on scalar arguments
- the integration algorithms and the 1D equation solvers should work on UniVariateFunction objects
- Point should *not* be persistent (maybe a new PersistentPoint class?) to avoid costly Counter creation