@jschueller 2019-09-03
yes, the idfactory definetely slows things down for us on such cases, and here again I was hoping my changes in AtomitInc using c++11's std::atomic would help a bit : |